# app/api/database-connections/[id]/analyze/route.ts
import { type NextRequest, NextResponse } from "next/server"
import { createClient } from "@/lib/supabase/server"
import { getCurrentUser } from "@/lib/auth-actions"

// POST /api/database-connections/[id]/analyze - Analisar banco de dados
export async function POST(request: NextRequest, { params }: { params: { id: string } }) {
  try {
    const user = await getCurrentUser()
    if (!user) {
      return NextResponse.json({ error: "Não autorizado" }, { status: 401 })
    }

    const supabase = await createClient()

    // Buscar conexão
    const { data: connection, error: connError } = await supabase
      .from("database_connections")
      .select("*")
      .eq("id", params.id)
      .eq("client_id", user.client_id)
      .single()

    if (connError || !connection) {
      return NextResponse.json({ error: "Conexão não encontrada" }, { status: 404 })
    }

    // Criar registro de análise
    const { data: analysis, error: analysisError } = await supabase
      .from("database_analyses")
      .insert({
        client_id: user.client_id,
        connection_id: params.id,
        status: "in_progress",
        started_at: new Date().toISOString(),
      })
      .select()
      .single()

    if (analysisError) throw analysisError

    // Aqui você implementaria a lógica real de análise do banco de dados
    // Por enquanto, vou simular uma análise básica

    // Simular descoberta de campos CNPJ
    const mockFindings = [
      {
        analysis_id: analysis.id,
        table_name: "empresas",
        column_name: "cnpj",
        data_type: "varchar(18)",
        finding_type: "cnpj_field",
        confidence: 0.95,
        sample_value: "12.345.678/0001-90",
      },
      {
        analysis_id: analysis.id,
        table_name: "fornecedores",
        column_name: "documento",
        data_type: "varchar(14)",
        finding_type: "cnpj_field",
        confidence: 0.85,
        sample_value: "98765432000100",
      },
    ]

    // Inserir findings
    await supabase.from("database_findings").insert(mockFindings)

    // Atualizar análise como concluída
    await supabase
      .from("database_analyses")
      .update({
        status: "completed",
        completed_at: new Date().toISOString(),
        total_tables: 2,
        total_fields_found: 2,
      })
      .eq("id", analysis.id)

    return NextResponse.json({
      analysis_id: analysis.id,
      message: "Análise iniciada com sucesso",
      findings_count: mockFindings.length,
    })
  } catch (error: any) {
    console.error("[v0] Error analyzing database:", error)
    return NextResponse.json({ error: error.message }, { status: 500 })
  }
}