# public/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CNPJ Alfanumérico API",
    "description": "API completa para análise e correção de campos CNPJ em repositórios de código e bancos de dados. Sistema multi-perfil com gestão de tarefas, integrações com Azure DevOps/GitHub e chatbot inteligente.",
    "version": "2.0.0",
    "contact": {
      "name": "Suporte CNPJ Alfanumérico",
      "email": "suporte@cnpj-alfa.com"
    },
    "license": {
      "name": "Proprietário",
      "url": "https://cnpj-alfa.com/license"
    }
  },
  "servers": [
    {
      "url": "https://api.cnpj-alfa.com",
      "description": "Servidor de Produção"
    },
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento Local"
    }
  ],
...
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "tags": ["Autenticação"],
        "summary": "Obter usuário atual",
        "description": "Retorna informações do usuário autenticado atualmente",
        "operationId": "getCurrentUser",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/clients": {
      "get": {
        "tags": ["Super Admin"],
        "summary": "Listar todos os clientes",
        "description": "Retorna lista completa de clientes cadastrados (apenas Super Admin)",
        "operationId": "listClients",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 20}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clients": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Client"}
                    },
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "totalPages": {"type": "integer"}
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Super Admin"],
        "summary": "Criar novo cliente",
        "description": "Cria um novo cliente no sistema com licença e configurações",
        "operationId": "createClient",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/chatbot": {
      "post": {
        "tags": ["Chatbot"],
        "summary": "Enviar mensagem ao chatbot",
        "description": "Interage com o chatbot inteligente para consultar tarefas e obter assistência",
        "operationId": "chatbotMessage",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["message", "user_id"],
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Minhas tarefas pendentes"
                  },
                  "user_id": {"type": "string", "format": "uuid"},
                  "session_id": {"type": "string", "format": "uuid"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do chatbot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reply": {"type": "string"},
                    "session_id": {"type": "string"},
                    "intent": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Chatbot"],
        "summary": "Obter histórico do chat",
        "description": "Retorna histórico de conversas do usuário",
        "operationId": "getChatHistory",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de conversas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "message": {"type": "string"},
                      "reply": {"type": "string"},
                      "timestamp": {"type": "string", "format": "date-time"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via endpoint /api/auth/login"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado - Token inválido ou ausente",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {"type": "string", "example": "Não autorizado"}
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Acesso negado - Permissão insuficiente",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {"type": "string", "example": "Acesso negado"}
              }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {"type": "string"},
                "details": {"type": "array", "items": {"type": "string"}}
              }
            }
          }
        }
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "email": {"type": "string", "format": "email"},
          "name": {"type": "string"},
          "role": {
            "type": "string",
            "enum": ["super_admin", "admin", "developer"]
          },
          "client_id": {"type": "string", "format": "uuid"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "Client": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "cnpj": {"type": "string"},
          "license_expires_at": {"type": "string", "format": "date-time"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateClientRequest": {
        "type": "object",
        "required": ["name", "cnpj", "admin_email", "admin_password"],
        "properties": {
          "name": {"type": "string", "example": "Empresa XYZ Ltda"},
          "cnpj": {"type": "string", "example": "12.345.678/0001-90"},
          "admin_email": {"type": "string", "format": "email"},
          "admin_password": {"type": "string", "format": "password"},
          "license_days": {"type": "integer", "default": 365}
        }
      },
...
    }
  },
  "security": [
    {"BearerAuth": []}
  ]
}